Map_manager, MapData, Map, Main_scene
Added a bunch of stuff for the map manager. Should be fully working with start and end points. Once the player reaches the end point they are teleported to the next level after 2 seconds. if they die they are reset to the levels starting position(could probably move the starting position to the start flag scene rather than be a part of the map).